dump

+++ Source/UserInterface/Views/NavigationDrawerDevider.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../JuceLibraryCode/JuceHeader.h"
+#include "../../../JuceLibraryCode/JuceHeader.h"
 #include "../../LookAndFeel/Material.h"
 
 class NavigationDrawerDevider : public Component {
@@ -17,7 +17,7 @@ public:
     void paint(Graphics &g) override {
         g.fillAll(Material::Color::Surface::Main);
         g.setColour(Material::Color::Icons::Selected::OnSurface::Active.withAlpha(0.12f));
-        g.drawLine(0, px(4), getWidth(), px(4));
+        g.drawLine(0.0f, px(4.0f), static_cast<float>(getWidth()), px(4.0f));
     }
 private:
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (NavigationDrawerDevider)
